initial organization and docs
