+++ HASP 24 Old/Digital.py
