+++ HASP 24 Old/TempSensors.py
