+++ HASP 24 Old/Utils.py
