+++ README.md
@@ -1,1 +1,3 @@
 # HASP-2025
+
+repository for COC's Hasp 2025 SpectraSolis Project.
